refactor(evaluation): extract repo file helpers
